Tests passing, but the HackathonSingle Snpashot does not display the content of the component
